requirements.txt
streamlit
cohere
sqlalchemy
psycopg2-binary
qdrant-client
python-dotenv
pandas
requests
dateparser
